Remove circular dependencies in react-router-dom
